Move KeyedLockStore into the RateLimit module
Kernel dissolution continues (#790): the generic keyed-lock primitive joins the
RateLimit module (it already depended on IntervalGate there — the module owns
the concurrency/throttling primitives, so the store's self-import disappears).

- KeyedLockStore -> Api/RateLimit/ (namespace Api.RateLimit); its test mirrors
  to Tests/RateLimit/.
- Reverse-edge check (playbook 3b): AvatarService takes per-user store locks
  through KeyedLockStore, so Avatar's DAG case gains the RateLimit edge
  (Avatar -> {Net, RateLimit}).

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1585 both TFMs, failures
checked by name in the same run's log: none), and the ABI-floor build.

Refs #790
